requirements.txt
Flask==2.2.5
pycryptodome==3.18.0
requests==2.31.0
streamlit==1.25.0
